Reworked Low Econ to Fix Defender AI problems
TLDR, we need to rework the resouce difficulty files as that was the culprit. I set low to vanilla values for now but we should rebalance the BOT values using vanilla as the basis for the MP, CP and defensive budget
The main issue was the the bot had a massive budget for defender AI so it would try to buy the "best" units it could
that's determined by the 'lvl' value in the settings_inf file in conquest. Example: {"mp/rus/mid/flamer"                    ("inf_tier3" side(rus) lvl(lvl2 lvl3 lvl4) r(1) rm(8) cost(90.0))}
reason flamers and AT were always selected is because they were some of the few inf that had lvl4 set (i.e. the best) and they are expensive
So we need to be much more mindful on how we scale defensive AI
set a min value for the budget so that we dont crash if the ai has too little

Other Changes:
- Re-enabled force AI cannon move to prevent AI cannons from idling after initial deployment on attack
- Reworked AI strategies to balance for the lower scale of low econ (1st pass)
- Slightly increased number of cloned AI defenders for player
- Fixed bug with radio scripts that enabled static wrecks to show on the minimap